Ship Shooting Overhaul
Created a base Shoot() method for the Ship class. Both the player and the NPC ships use this method create projectiles based on the slotted weapon scriptable object.
This allows for a very "Plug and play" approach to enemies, where if you want to swap weapons, literally all you have to do is change the scriptable object and the code will handle the rest.