Projectile improvements
Improved the projectiles to use values from the scriptable object Weapon that is used to initialize them, instead of just using whatever the prefab says.

This allows for the one projectile prefab to be used for many kinds of weapons. However, weapons that have different behaviors (such as a very long laser, for example, would need different prefabs)