Boss Basics
The boss currently functions exactly like a normal ship, except that it has many other ships nested underneath it. These ships are "Turrets".

There is currently a very annoying bug that results in the bars of the turrets snapping into place when they should not be. There is an easy fix for this, but I do not like the solution since it is a little bit messy.