Movement speed consistency fix
Changed the force addition to compensate for mass and drag... This way the "speed" number that is edited in the window actually represents the speed at which the ships will go.

Example: A ship with mass of 5000 and drag of 5 with speed 10 moves as fast as a ship with mass of 100 and drag of 1.