Renamed Ship.cs -> to ShipController.cs
Did this because we will probably want to create some sort of actual representation of a "Ship" in game.

Likely a scriptable object that contains generic information such as the image of the ship, animator controller, some basic stats.

It would make for sense for this container to be called "Ship" instead of something like "ShipContainer", and the script that is now ShipController matches all of the other scripts such as "TurretController", "PlayerController", "NpcController", "BossController"